// types.ts
export type ModulePageContentType =
  | 'title'
  | 'paragraph'
  | 'list'
  | 'quote'
  | 'image'
  | 'video'
  | 'divider'
  | 'subtitle';

export interface ModulePageContent {
  type: ModulePageContentType;
  content: string | string[];
  alt?: string; // For images
  align?: 'left' | 'center' | 'right' | 'justify';
  color?: string; // For text color
}

export interface ModulePage {
  id: number;
  title: string;
  content: ModulePageContent[];
}